pyproject.toml
[project]
name = "sto-link-image-backend"
version = "1.0.0"
description = "StoLink Image Worker - AWS Bedrock based image generation service"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic-settings>=2.6.0",
    "aio-pika>=9.5.0",
    "boto3>=1.35.0",
    "requests>=2.32.0",
    "httpx>=0.28.0",
    # LangGraph & LangChain
    "langgraph>=0.2.0",
    "langchain-aws>=0.2.0",
    "langchain-core>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
